backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
motor==3.2.0
pydantic==2.3.0
pydantic[email]==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
python-dotenv==1.0.0
httpx==0.24.1
jinja2==3.1.2